@@ -1,73 +1,85 @@
 {
   "name": "ts-cleanup",
-  "version": "0.2.6",
   "description": "A tool for cleaning up files built by typeScript",
+  "version": "0.2.6",
+  "author": "Panates <info@panates.com>",
+  "license": "MIT",
+  "private": true,
+  "type": "module",
+  "main": "./cjs/index.js",
+  "module": "./esm/index.js",
+  "types": "typings/index.d.ts",
+  "dependencies": {
+    "chokidar": "^3.6.0",
+    "colors": "^1.4.0",
+    "commander": "^12.1.0",
+    "fast-glob": "^3.3.2"
+  },
+  "devDependencies": {
+    "@panates/eslint-config": "^1.0.3",
+    "@panates/eslint-config-ts": "^1.0.3",
+    "@panates/tsconfig": "^1.0.3",
+    "@types/chokidar": "^2.1.3",
+    "@types/colors": "^1.2.1",
+    "@types/commander": "^2.12.2",
+    "@types/jest": "^29.5.12",
+    "@types/node": "^20.14.9",
+    "@types/temp": "^0.9.4",
+    "jest": "^29.7.0",
+    "prettier": "^3.3.2",
+    "temp": "^0.9.4",
+    "ts-jest": "^29.1.5",
+    "ts-loader": "^9.5.1",
+    "ts-node": "^10.9.2",
+    "typescript": "5.4.5"
+  },
+  "scripts": {
+    "clean": "tsc && npm run clean:test && npm run clean:src",
+    "clean:src": "node ./dist/bin -s src -d dist -a",
+    "clean:test": "node ./dist/bin -s test",
+    "build": "npm run build:cjs && npm run build:esm",
+    "build:cjs": "tsc -b tsconfig-build-cjs.json",
+    "build:esm": "tsc -b tsconfig-build-esm.json",
+    "postbuild": "npm run postbuild:copyfiles && node ./support/postbuild.cjs",
+    "postbuild:copyfiles": "cp support/package.cjs.json ./build/cjs/package.json && cp LICENSE README.md ./build && cp -r ./bin ./build",
+    "lint:src": "tslint -p tsconfig.json \"src/**/*.ts\" --fix",
+    "lint:spec": "tslint -p tsconfig.spec.json \"test/**/*.spec.ts\" --fix",
+    "test": "jest",
+    "lint": "eslint . --max-warnings=0",
+    "lint:fix": "eslint . --max-warnings=0 --fix",
+    "format": "prettier . --write --log-level=warn",
+    "precover": "rimraf coverage",
+    "cover": "jest --runInBand --detectOpenHandles --coverage",
+    "precitest": "rimraf coverage",
+    "citest": "jest --coverage --coverageReporters=lcov"
+  },
+  "bin": {
+    "ts-cleanup": "bin/ts-cleanup.mjs"
+  },
+  "files": [
+    "bin/",
+    "cjs/",
+    "esm/",
+    "typings/",
+    "LICENSE",
+    "README.md",
+    "CHANGELOG.md"
+  ],
+  "bugs": {
+    "url": "https://github.com/panates/ts-cleanup.git/issues"
+  },
+  "engines": {
+    "node": ">= 16.0"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/panates/ts-cleanup.git"
+  },
   "keywords": [
     "ts",
     "typescript",
     "clean",
     "cleanup",
     "tidy"
-  ],
-  "author": "Panates <info@panates.com>",
-  "license": "MIT",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "directories": {
-    "lib": "./dist"
-  },
-  "bin": {
-    "ts-cleanup": "dist/bin.js"
-  },
-  "dependencies": {
-    "chokidar": "^3.4.2",
-    "colors": "^1.4.0",
-    "commander": "^9.5.0",
-    "fast-glob": "^3.2.12"
-  },
-  "devDependencies": {
-    "@types/chokidar": "^2.1.3",
-    "@types/colors": "^1.2.1",
-    "@types/commander": "^2.12.2",
-    "@types/mocha": "^10.0.1",
-    "@types/node": "^18.11.18",
-    "mocha": "^10.2.0",
-    "nyc": "^15.1.0",
-    "temp": "^0.9.4",
-    "ts-loader": "^9.4.2",
-    "ts-node": "^10.9.1",
-    "tslint": "^6.1.3",
-    "typescript": "^4.9.4"
-  },
-  "files": [
-    "LICENSE",
-    "README",
-    "dist/"
-  ],
-  "engines": {
-    "node": ">= 8.0"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/panates/ts-cleanup.git"
-  },
-  "scripts": {
-    "test": "mocha -r ts-node/register --reporter spec test/**/*.spec.ts",
-    "cover": "nyc --reporter=cobertura --reporter html --reporter text npm run test",
-    "travis-cover": "nyc --reporter lcovonly npm run test",
-    "clean": "tsc && npm run clean:test && npm run clean:src",
-    "clean:src": "node ./dist/bin -s src -d dist -a",
-    "clean:test": "node ./dist/bin -s test",
-    "preBuild": "npm run clean",
-    "build": "npm run clean && npm run lint && tsc",
-    "lint": "npm run lint:src && npm run lint:spec",
-    "lint:src": "tslint -p tsconfig.json \"src/**/*.ts\" --fix",
-    "lint:spec": "tslint -p tsconfig.spec.json \"test/**/*.spec.ts\" --fix"
-  },
-  "bugs": {
-    "url": "https://github.com/panates/ts-cleanup.git/issues"
-  },
-  "nyc": {
-    "temp-dir": "./coverage/.nyc_output"
-  }
+  ]
 }
